Feat: separating codes related to implementing linked list from rw_semaphore_module.c
